pyproject.toml
[tool.poetry.dependencies]
python = ">=3.12,<4.0"
zeep = "4.2.1"
pydantic = "1.10.8"
simhash = "2.1.2"
loguru = "0.7.0"
pandas = "2.0.1"
openpyxl = "3.1.2"
python-dotenv = "1.0.0"
tzdata = "*"
requests = "2.31.0"
# Добавленные веб-зависимости
flask = "2.2.3"
flask-login = "0.6.2"
flask-wtf = "1.1.1"
flask-sqlalchemy = "3.0.3"
flask-migrate = "4.0.4"
werkzeug = "2.2.3"
celery = "5.2.7"
redis = "4.5.4"
gunicorn = "20.1.0"
# Дополнительные зависимости
jinja2 = "3.1.2"
itsdangerous = "2.1.2"
click = "8.1.3"
markupsafe = "2.1.2"
email-validator = "1.3.1"  # Для проверки email